no need for grDevices:: before dev.hold/dev.flush
Typically we'd use <package>:: when the package is not included
a bunch or is from base package lists. It's useful to be specific
from which package we're taking a function from in the case that
it's not in base packages (or not used super-widely in the code)
so that you make sure you get the right function.

For stuff like spatstat we wouldn't, as we use it a LOT, so that
would mean a whole heap of spatstat:: stuff for no reason. For
animation we probably would though (should that become a
dependency at a later date).