Ignore kernels more than 6 bandwidths away from the evaluation point.
This gives identical estimates on the PBC data, but reduces the number
of kernel evaluations by 2.

It is plausible that binning could be done here if we feel it is useful
to save the exp(). Am not sure what the expense of exp is and whether
it is worth doing it. Presumably it is significant multiplies/adds??